very small change. it spreads input styles to `css()`, improves glamor's numbers by ~20%