@threepointone 2017-04-07
added another fix, shaving an additional 20-30% off glamor's numbers.