@threepointone 2017-04-07
before - 
```
performance.bundle.js:36 Deep tree [glamor]
178.19ms ±10.84ms
performance.bundle.js:36 Wide tree [glamor]
311.17ms ±13.76ms
```

after -
```
performance.bundle.js:36 Deep tree [glamor]
83.67ms ±11.46ms
performance.bundle.js:36 Wide tree [glamor]
164.42ms ±9.98ms
```