@necolas 2017-04-08
Thanks! I figured I'd done something wrong as there's no way glamor should be almost as slow as styled-components