@necolas 2017-04-09
Applied in https://github.com/necolas/react-native-web/commit/aa85876eb20b56cc086780499b52d737586c43e4