@threepointone 2017-04-09
thanks much! it should also get a touch faster with the next release of glamor; with that I'm currently getting ~10ms diff with rnw on deep tree, and ~20-25ms on the wide tree. 

thanks for pushing me to do this!